upgrade settings to AppCompatActivity (remove tablet view) / streamline preference access
