--- News-Android-App/src/main/res/xml/pref_headers.xml
@@ -1,50 +1,0 @@
-<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <!-- These settings headers are only used on tablets. -->
-
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$EmptyPreferenceFragment"
-        android:title=""
-        app:iconSpaceReserved="false"/>
-
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$GeneralPreferenceFragment"
-        android:title="@string/pref_header_general"
-        app:iconSpaceReserved="false"/>
-
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$DisplayPreferenceFragment"
-        android:title="@string/pref_header_display"
-        app:iconSpaceReserved="false"/>
-
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$NotificationPreferenceFragment"
-        android:title="@string/pref_header_notifications"
-        app:iconSpaceReserved="false"/>
-
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$DataSyncPreferenceFragment"
-        android:title="@string/pref_header_data_sync"
-        app:iconSpaceReserved="false"/>
-
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$AboutPreferenceFragment"
-        android:title="@string/pref_header_about"
-        app:iconSpaceReserved="false"/>
-
-
-    <!--
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$PodcastPreferenceFragment"
-        android:title="@string/pref_header_podcast"
-        app:iconSpaceReserved="false"/>
-    -->
-
-    <!--
-    <Preference
-        android:fragment="de.luhmer.owncloudnewsreader.SettingsActivity$DataSyncPreferenceFragment"
-        android:title="@string/pref_header_data_sync"
-        app:iconSpaceReserved="false"/>
--->
-</androidx.preference.PreferenceScreen>
